backend/requirements.txt
# =============================================================================
# Atlas RAG System - Python Dependencies
# =============================================================================

# Web Framework
fastapi==0.115.6
uvicorn[standard]==0.34.0
python-multipart==0.0.20

# Database
sqlalchemy==2.0.36
asyncpg==0.30.0
pgvector==0.3.6

# Authentication
pyjwt==2.10.1
bcrypt==4.2.1

# Configuration
pyyaml==6.0.2
pydantic==2.10.4

# HTTP Client (for Docling API and vLLM calls)
httpx==0.28.1

# PDF page rendering (for page-view endpoint)
pypdf==5.1.0
pdf2image==1.17.0
Pillow==11.1.0

# Cross-encoder reranking (ONNX, no PyTorch needed)
flashrank

# Docker
docker==7.1.0

# Testing
pytest==8.3.4
pytest-asyncio==0.25.0
